[CJ/EC][76395676] stronger pred for child isa? env
Discovered when tests were passing which should have failed.
Passing silently, if you will.

The former condition was ≈ (or (env? k) (coll? k))
which really wasn't specific enough.

What we really want is in the code.
(if k is a collection, it must contain an env or default)

Thanks for listening.